Vector Coords are NOW doubles

+++ source/environment/MathVector.hpp
@@ -12,11 +12,11 @@
 // WILL BE REPLACED WHEN OPENGL IS IMPLEMENTED
 class Vector2d{
 	public:
-		float x;
-		float y;
+		double x;
+		double y;
 		Vector2d(){x = 0; y = 0;}
 		Vector2d(int n_x, int n_y){x = n_x; y = n_y;}
-		float dot(Vector2d other) {return (x * other.x + y * other.y);} 
+		double dot(Vector2d other) {return (x * other.x + y * other.y);} 
 };
 
 #endif
